refactor: split flex-toggle-group.tsx
